Moved process_unit_bearing_argument() from prepare.py to utils.py since it will be used during YAML parsing
