+++ Yank/utils.py
@@ -4,6 +4,8 @@ import collections
 
 from pkg_resources import resource_filename
 
+from simtk import unit
+
 #========================================================================================
 # Logging functions
 #========================================================================================
@@ -224,6 +226,44 @@ def get_data_filename(relative_path):
         raise ValueError("Sorry! %s does not exist. If you just added it, you'll have to re-install" % fn)
 
     return fn
+
+def process_unit_bearing_argument(args, argname, compatible_units):
+    """
+    Process a unit-bearing command-line argument to produce a Quantity.
+
+    Parameters
+    ----------
+    args : dict
+       Command-line arguments dict from docopt.
+    argname : str
+       Key to use to extract value from args.
+    compatible_units : simtk.unit.Unit
+       The result will be checked for compatibility with specified units, and an exception raised if not compatible.
+
+    Returns
+    -------
+    quantity : simtk.unit.Quantity
+       The specified parameter, returned as a Quantity.
+
+    """
+
+    # WARNING: This is dangerous!
+    # See: http://nedbatchelder.com/blog/201206/eval_really_is_dangerous.html
+    # TODO: Can we use a safer form of (or alternative to) 'eval' here?
+    quantity = eval(args[argname], unit.__dict__)
+    # Unpack quantity if it was surrounded by quotes.
+    if isinstance(quantity, str):
+        quantity = eval(quantity, unit.__dict__)
+    # Check to make sure units are compatible with expected units.
+    try:
+        quantity.unit.is_compatible(compatible_units)
+    except:
+        raise Exception("Argument %s does not have units attached." % args[argname])
+    # Check that units are compatible with what we expect.
+    if not quantity.unit.is_compatible(compatible_units):
+        raise Exception("Argument %s must be compatible with units %s" % (argname, str(compatible_units)))
+    # Return unit-bearing quantity.
+    return quantity
 
 #=============================================================================================
 # Main and tests
